Added Assets Page

+++ src/client-portal/Client.Infrastucture/Extensions/NumberExtensions.cs
@@ -6,24 +6,24 @@ namespace Client.Infrastructure.Extensions
 {
     public static class NumberExtensions
     {
-        public static decimal ToAmount(this BigInteger value, int decimals)
+        public static decimal ToAmount(this long value, int decimals)
         {
             return Convert.ToDecimal(((double)value) / Math.Pow(10, decimals));
         }
 
-        public static decimal ToAmount(this BigInteger value, byte decimals)
+        public static decimal ToAmount(this long value, byte decimals)
         {
             return Convert.ToDecimal(((double)value) / Math.Pow(10, decimals));
         }
 
         public static BigInteger ToAmount(this decimal value, int decimals)
         {
-            return (BigInteger)((double)value * Math.Pow(10, decimals));
+            return (long)((double)value * Math.Pow(10, decimals));
         }
 
         public static BigInteger ToAmount(this double value, int decimals)
         {
-            return (BigInteger)(value * Math.Pow(10, decimals));
+            return (long)(value * Math.Pow(10, decimals));
         }
 
         public static int CountDecimalPlaces(this double value)
